Okay so old atari-py doesn't have some of the games (e.g. hero) and anyway I'm not sure yet whether the black I'm seeing is just air-raid or if other games have it too. So...running other games.

+++ continual_rl/policies/sane/hypothesis/hypothesis.py
@@ -14,7 +14,7 @@ class InputScaler(nn.Module):
         self._observation_space = observation_space
 
     def forward(self, x):
-        return x/self._observation_space.high
+        return x.float()/self._observation_space.high
 
 
 class ConvNetTimeWrapper(nn.Module):
